refactor: BaseTelegramCommand.command_text is now command
The type of this attribute also changed from string to type[BaseBotCommand] | None.
The bot will still raise an error if the command was not provided, but we no longer
need to hardcode the commands. This is more user-friendly as a refactor of the command
would change the import path and so IDEs would automatically update them. While plain
strings would never be updated and easily missed.