feat: do not let TelegramSettings assume a user link
Since this is a re-usable library, it's best to keep this as light
and 'dumb' as possible. Initially it was assumed that a user will be
present on TelegramSettings, but this is not always the case.
So we no longer depend on a user, but allow more customization where
needed