fix: do not fetch callbackdata in Command.previous_step
Before this commit the _steps_back were configured on the callbackdata,
but this is not needed as we have access to the steps in the command,
so also to the steps_back attribute. This also means users can no
longer hack around steps_back by providing _steps_back in the kwargs manually.
If this need is ever there, an issue should be logged for it.

In this commit, the text shown on the back button is also made configurable.