feat: add translation support
Command.translate decides whether or not the steps should be translated.
Step.translate is None by default, meaning the command's setting will be
used. When Step.translate is True or False, this value will be used.
Step.__call__ is now used as an entrypoint to start the step. It essentially
just activates translation and calls Step.handle.

Help messages are always translated, it's currently not supported to disable
translation on this.